requirements.txt
dotted-dict
kubernetes
mergedeep
requests
ruamel.yaml
